Header was rewrited to view
