+++ js/screens/game-screen.js
@@ -1,14 +1,13 @@
 import {GameParam} from '../data/constants';
-import LevelTypeOne from '../view/level-type-1-view';
-import LevelTypeTwo from '../view/level-type-2-view';
-import LevelTypeThree from '../view/level-type-3-view';
+import LevelTypeOne from '../views/level-type-1-view';
+import LevelTypeTwo from '../views/level-type-2-view';
+import LevelTypeThree from '../views/level-type-3-view';
 import Application from '../application';
 
 export default class GameScreen {
   constructor(model) {
     this.model = model;
     this._content = new LevelTypeOne(this.model._gameState);
-    this.timerContainer = this._content.element.querySelector(`.game__timer`);
     this._timer = null;
   }
 
@@ -30,7 +29,7 @@ export default class GameScreen {
   }
 
   updateHeader() {
-    this.timerContainer.innerHTML = this.model._gameState.timer;
+
   }
 
   startGame() {
